backend/requirements.txt
# Core FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
supabase==2.10.0

# OCR & Image Processing (Fixed versions for Python 3.13 compatibility)
# Install paddlepaddle separately first: pip install paddlepaddle==3.0.0b1 -i https://www.paddlepaddle.org.cn/packages/stable/cpu/
paddleocr==2.9.1
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
aiofiles==24.1.0

bcrypt==4.0.1
passlib==1.7.4

groq==0.2.1

# ML (optional for now - comment out if causing issues)
# ultralytics==8.3.28
# scikit-learn==1.5.2